# swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "PrsApi",
        "version": "v1"
    },
    "paths": {
        "/api/Products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/Products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/PurchaseRequests": {
            "get": {
                "tags": [
                    "PurchaseRequests"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PurchaseRequest"
                                    }
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PurchaseRequest"
                                    }
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PurchaseRequest"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "PurchaseRequests"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PurchaseRequest"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PurchaseRequest"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/PurchaseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/PurchaseRequest"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PurchaseRequest"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PurchaseRequest"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/PurchaseRequests/{id}": {
            "get": {
                "tags": [
                    "PurchaseRequests"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/PurchaseRequest"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PurchaseRequest"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PurchaseRequest"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "PurchaseRequests"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PurchaseRequest"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PurchaseRequest"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/PurchaseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "PurchaseRequests"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "required": [
                    "name",
                    "price",
                    "unit",
                    "vendorId",
                    "vendorPartNumber"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "vendorPartNumber": {
                        "maxLength": 30,
                        "minLength": 0,
                        "type": "string"
                    },
                    "name": {
                        "maxLength": 30,
                        "minLength": 0,
                        "type": "string"
                    },
                    "price": {
                        "type": "number",
                        "format": "double"
                    },
                    "unit": {
                        "maxLength": 30,
                        "minLength": 0,
                        "type": "string"
                    },
                    "photopath": {
                        "maxLength": 255,
                        "minLength": 0,
                        "type": "string",
                        "nullable": true
                    },
                    "vendorId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "vendorNavigation": {
                        "$ref": "#/components/schemas/Vendor"
                    }
                },
                "additionalProperties": false
            },
            "PurchaseRequest": {
                "required": [
                    "deliveryMode",
                    "description",
                    "justification",
                    "status"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "description": {
                        "maxLength": 80,
                        "minLength": 0,
                        "type": "string"
                    },
                    "justification": {
                        "maxLength": 80,
                        "minLength": 0,
                        "type": "string"
                    },
                    "reasonForRejection": {
                        "maxLength": 80,
                        "minLength": 0,
                        "type": "string",
                        "nullable": true
                    },
                    "dateNeeded": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "deliveryMode": {
                        "maxLength": 20,
                        "minLength": 0,
                        "type": "string"
                    },
                    "status": {
                        "maxLength": 10,
                        "minLength": 0,
                        "type": "string"
                    },
                    "total": {
                        "type": "number",
                        "format": "double"
                    },
                    "submittedDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "userId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "userNavigation": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "additionalProperties": false
            },
...
            "User": {
                "required": [
                    "firstName",
                    "isAdmin",
                    "isReviewer",
                    "lastName",
                    "password",
                    "userName"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "userName": {
                        "maxLength": 30,
                        "minLength": 0,
                        "type": "string"
                    },
                    "password": {
                        "maxLength": 30,
                        "minLength": 0,
                        "type": "string"
                    },
                    "firstName": {
                        "maxLength": 30,
                        "minLength": 0,
                        "type": "string"
                    },
                    "lastName": {
                        "maxLength": 30,
                        "minLength": 0,
                        "type": "string"
                    },
                    "phone": {
                        "maxLength": 12,
                        "minLength": 0,
                        "type": "string",
                        "nullable": true
                    },
                    "email": {
                        "maxLength": 255,
                        "minLength": 0,
                        "type": "string",
                        "nullable": true
                    },
                    "isAdmin": {
                        "type": "boolean"
                    },
                    "isReviewer": {
                        "type": "boolean"
                    }
                },
                "additionalProperties": false
            },
            "Vendor": {
                "required": [
                    "address",
                    "city",
                    "code",
                    "email",
                    "name",
                    "phone",
                    "state",
                    "zip"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "code": {
                        "maxLength": 30,
                        "minLength": 0,
                        "type": "string"
                    },
                    "name": {
                        "maxLength": 30,
                        "minLength": 0,
                        "type": "string"
                    },
                    "address": {
                        "maxLength": 30,
                        "minLength": 0,
                        "type": "string"
                    },
                    "city": {
                        "maxLength": 30,
                        "minLength": 0,
                        "type": "string"
                    },
                    "state": {
                        "maxLength": 2,
                        "minLength": 0,
                        "type": "string"
                    },
                    "zip": {
                        "maxLength": 5,
                        "minLength": 0,
                        "type": "string"
                    },
                    "phone": {
                        "maxLength": 12,
                        "minLength": 0,
                        "type": "string"
                    },
                    "email": {
                        "maxLength": 255,
                        "minLength": 0,
                        "type": "string"
                    }
                },
                "additionalProperties": false
            },
...
        },
        "securitySchemes": {
            "oauth2": {
                "type": "oauth2",
                "flows": {
                    "authorizationCode": {
                        "authorizationUrl": "http://localhost:8080/realms/bootcamp/protocol/openid-connect/auth",
                        "tokenUrl": "http://localhost:8080/realms/bootcamp/protocol/openid-connect/token",
                        "scopes": {
                            "openid": "OpenID Connect scope",
                            "profile": "Profile scope",
                            "email": "Email scope"
                        }
                    }
                }
            }
        }
    },
    "security": [
        {
            "oauth2": [
                "openid",
                "profile",
                "email"
            ]
        }
    ]
}